backend/requirements.txt
Flask==2.0.2
Flask-CORS==5.0.1
Flask-JWT-Extended==4.4.4
requests==2.26.0
bcrypt==3.2.0
Werkzeug==2.0.3
python-dotenv